% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_onsets.R
\name{align_onsets}
\alias{align_onsets}
\title{Align to stimulus onset
TODO: base alignment on mean of region 66 & 67}
\usage{
align_onsets(
  data,
  align_region = 66,
  frame_filter = 15,
  bg = 1:8,
  sd_multi = 4,
  onset_intended = 10,
  cutframes = -5:45,
  value_column = "dff"
)
}
\arguments{
\item{data}{dataframe (of dff data, for now (data_t))}

\item{align_region}{region to align to (usually sensilla input region)}

\item{frame_filter}{max frame to look for onset}

\item{bg}{background frames for noise SD level calculation}

\item{sd_multi}{value to multiply bg sd by to define onset detection threshold}

\item{onset_intended}{intended stimulus onset (to shift nostim and when no response is detected)}

\item{cutframes}{frames to reurn after shifting}

\item{value_column}{name of the column that contains the values}
}
\value{

}
\description{
Align to stimulus onset
TODO: base alignment on mean of region 66 & 67
}
