% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_ts.R
\name{concatenate_ts}
\alias{concatenate_ts}
\title{concatenate TS
concatenate time series data by changing the frame}
\usage{
concatenate_ts(data, include_nostim = FALSE)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
concatenate TS
concatenate time series data by changing the frame
}
