% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate_ts2.R
\name{concatenate_ts2}
\alias{concatenate_ts2}
\title{concatenate TS
concatenate time series data by changing the frame}
\usage{
concatenate_ts2(
  data,
  stimuli = c("nostim 0", "h2o 0", "sucrose 200mM", "yeast 10\%")
)
}
\arguments{
\item{data}{}

\item{stimuli}{stimnuli in required order}
}
\value{

}
\description{
concatenate TS
concatenate time series data by changing the frame
}
