% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_regions.R
\name{filter_regions}
\alias{filter_regions}
\title{filter_region
Filter regions based on: "response probablility"}
\usage{
filter_regions(data, method = "P", SD = 4, P = 0.5, bg = -5:0, res = 1:5)
}
\arguments{
\item{data}{DFF time series dataframe}

\item{method}{filtering method. P == (response probability) or R for responses above the SD rersponse threshold}

\item{SD}{multiplicator. How many times pre-stimulus SD to count as a response?}

\item{P}{probability threshold for filter.}

\item{bg}{Frames for calculating background SD.}

\item{res}{Frames of response time window.}
}
\value{
vector of regions that pass filter in at least one internal state
}
\description{
filter_region
Filter regions based on: "response probablility"
}
