% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_onsets.R
\name{find_onsets}
\alias{find_onsets}
\title{Find stimulus onsets}
\usage{
find_onsets(data, frame_filter = 30, bg = -5:-1, sd_multi = 4)
}
\arguments{
\item{data}{dataframe (of dff data, for now (data_t))}

\item{frame_filter}{max frame to look for onset}

\item{bg}{background frames for noise SD level calculation}

\item{sd_multi}{value to multiply bg sd by to define onset detection threshold}
}
\value{

}
\description{
Find stimulus onsets
}
