% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_response.R
\name{peak_response}
\alias{peak_response}
\title{Calculate peak responses}
\usage{
peak_response(data, method = "mean", frames = 1:5, bgf = -5:0, sd_mp = 2)
}
\arguments{
\item{data}{dff data (TODO: generalize to other data)}

\item{method}{calculation method, defaults to "mean" of ´frames´}

\item{frames}{frames to perform the calculation on}

\item{bgf}{background frames}

\item{sd_mp}{multiplyer for SD threshold response detection (peak > sd_bg * sd_mp)
use same parameters as in `filter_regions`.}
}
\value{
data frame containing peak values
}
\description{
Calculate peak responses
}
