% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_atlas.R
\name{plot_atlas}
\alias{plot_atlas}
\title{Plot data onto atlas representation}
\usage{
plot_atlas(
  data = NULL,
  atlas = sez_atlas,
  value_column = "value",
  region_column = "region",
  value_name = "value",
  plot_legend = FALSE,
  fill_color = "grey70",
  line_color = "grey30",
  alpha = 1,
  line_width = 0.2,
  label = FALSE,
  label_color = "grey20",
  label_size = 2
)
}
\arguments{
\item{line_color}{line color, set to "MAP" in order to map to the value column}

\item{alpha}{alpha value of the polygons}

\item{label_size}{}
}
\value{

}
\description{
Plot data onto atlas representation
}
