% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_helpers.R
\name{theme_dahaniel2}
\alias{theme_dahaniel2}
\title{Dahaniel custom ggplot theme}
\usage{
theme_dahaniel2(
  base_size = 5,
  base_family = "Arial",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{base_family}{}
}
\value{

}
\description{
Dahaniel custom ggplot theme
}
